Created prefabs, began work on level design for level 2, updated EnemySight script.
Made boxes and the container into prefabs, for ease of use later.

Began level design for level 2, adding obstructions as well as enemies on to the level. Level design has be made up onto the first Radiostation.

Updated the movabe EnemySight script, so that the enemies have a max range in which they can communicate with eachother.