Added new weapon. Updated rifle look. Added audio. Continued work on fleshing out the tunnels.
Added a new secondary weapon, a silenced pistol.
The silenced pistol does 10x more damage to a moveable enemy that is not in CombatState.Combat, and 10x less damage to one that is.

I also added weapon swapping, as well as doing some cleanup in the code for the rifle, and now also pistol, scripts for when they move around (i.e. aiming and also now holstering)

I updated to look of the rifle, so it now uses the Tar21 from the POLYGON weapons pack.

Added audio for the silencer shooting, as well as for the Radiostation explosion.

Did some more work on the tunnels.