Updated PlayerMovement script, Pistol script, and Rifle script.
Updated PlayerMovement.Update() section, which handles weapon switching. It is now possible to switch weapons using either the mouse scroll wheel, or the keyboard buttons, 1 and 2.

Updated Pistol.reloadCooldown to make it shorter, so it better fits with the audio clip.

Update Rifle.reloadCooldown to make it shorter, so it better fits with the audio clip.